stores: move migrations code into separate file; create index using migrator instead of raw sql
